minor bug fix (#3)
* added random clipping to clip_signal

added a random cropping to clip_signal utility function, this will crop the signa starting at a random location such that the cropped signal is duration-second-long.

* added docstring for clip_signal

self-explanatory really

* forgot to stage

* added random cropping to GTZAN dataset

added parameter `random_crops` which specify how many crops from an audio to create, where the start position is randomized for every crops.

* renamed clip_signal to crop_signal

* updated training script configuration

- refactor the layout of the train_config.yml file and added a new argument `random_crops`.
- reflect changes to `input.py`, `preparations.py` and `train.py`

* bugfix: random split bug fix

If random_crops is given, the random split will introduce a data leak, so a manual random split must be done on the files instead of the indices of the dataset.

* minor seeding update

Seed the random module as well because it is used in data_utils/processing_utils.py